+++ cada/models.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 from flask_mongoengine import MongoEngine
 
 db = MongoEngine()