+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 - Add a `burnthemall` command [#25](https://github.com/etalab/cada.data.gouv.fr/pull/25)
 - Use `tqdm` for major commands with loops [#25](https://github.com/etalab/cada.data.gouv.fr/pull/25)
+- Python 3 [#28](https://github.com/etalab/cada.data.gouv.fr/pull/28)
 - Elastic Search 7.2 [#29](https://github.com/etalab/cada.data.gouv.fr/pull/29)
 - Mongo 4.1 [#29](https://github.com/etalab/cada.data.gouv.fr/pull/29)
-- Python 3 [#28](https://github.com/etalab/cada.data.gouv.fr/pull/28)
+- Improved pseudonymization [#32](https://github.com/etalab/cada.data.gouv.fr/pull/32)
+- Display sample of api calls [#33](https://github.com/etalab/cada.data.gouv.fr/pull/33)
+- Fix text query [#33](https://github.com/etalab/cada.data.gouv.fr/pull/33)
+
 
 ## 0.2.1 (2019-07-15)